functional access to system table
